:zap:add es7

+++ openjob-server/openjob-server-admin/src/main/java/io/openjob/server/admin/request/job/ListProcessorLogRequest.java
@@ -37,5 +37,5 @@ public class ListProcessorLogRequest {
     private Long time = 0L;
 
     @ApiModelProperty(value = "Page size", required = true)
-    private Long size = 20L;
+    private Integer size = 20;
 }
